+++ data/hydration-test-data.js
@@ -1,0 +1,761 @@
+const hydrationData = [
+  {
+    "userID": 1,
+    "date": "2019/06/15",
+    "numOunces": 37
+  },
+  {
+    "userID": 2,
+    "date": "2019/06/15",
+    "numOunces": 75
+  },
+  {
+    "userID": 3,
+    "date": "2019/06/15",
+    "numOunces": 47
+  },
+  {
+    "userID": 4,
+    "date": "2019/06/15",
+    "numOunces": 85
+  },
+  {
+    "userID": 5,
+    "date": "2019/06/15",
+    "numOunces": 42
+  },
+  {
+    "userID": 6,
+    "date": "2019/06/15",
+    "numOunces": 87
+  },
+  {
+    "userID": 7,
+    "date": "2019/06/15",
+    "numOunces": 94
+  },
+  {
+    "userID": 8,
+    "date": "2019/06/15",
+    "numOunces": 84
+  },
+  {
+    "userID": 9,
+    "date": "2019/06/15",
+    "numOunces": 39
+  },
+  {
+    "userID": 10,
+    "date": "2019/06/15",
+    "numOunces": 75
+  },
+  {
+    "userID": 11,
+    "date": "2019/06/15",
+    "numOunces": 51
+  },
+  {
+    "userID": 1,
+    "date": "2019/06/16",
+    "numOunces": 69
+  },
+  {
+    "userID": 2,
+    "date": "2019/06/16",
+    "numOunces": 91
+  },
+  {
+    "userID": 3,
+    "date": "2019/06/16",
+    "numOunces": 99
+  },
+  {
+    "userID": 4,
+    "date": "2019/06/16",
+    "numOunces": 95
+  },
+  {
+    "userID": 5,
+    "date": "2019/06/16",
+    "numOunces": 79
+  },
+  {
+    "userID": 6,
+    "date": "2019/06/16",
+    "numOunces": 57
+  },
+  {
+    "userID": 7,
+    "date": "2019/06/16",
+    "numOunces": 49
+  },
+  {
+    "userID": 8,
+    "date": "2019/06/16",
+    "numOunces": 26
+  },
+  {
+    "userID": 9,
+    "date": "2019/06/16",
+    "numOunces": 98
+  },
+  {
+    "userID": 10,
+    "date": "2019/06/16",
+    "numOunces": 68
+  },
+  {
+    "userID": 1,
+    "date": "2019/06/17",
+    "numOunces": 96
+  },
+  {
+    "userID": 2,
+    "date": "2019/06/17",
+    "numOunces": 96
+  },
+  {
+    "userID": 3,
+    "date": "2019/06/17",
+    "numOunces": 28
+  },
+  {
+    "userID": 4,
+    "date": "2019/06/17",
+    "numOunces": 82
+  },
+  {
+    "userID": 5,
+    "date": "2019/06/17",
+    "numOunces": 99
+  },
+  {
+    "userID": 6,
+    "date": "2019/06/17",
+    "numOunces": 57
+  },
+  {
+    "userID": 7,
+    "date": "2019/06/17",
+    "numOunces": 50
+  },
+  {
+    "userID": 8,
+    "date": "2019/06/17",
+    "numOunces": 47
+  },
+  {
+    "userID": 9,
+    "date": "2019/06/17",
+    "numOunces": 86
+  },
+  {
+    "userID": 10,
+    "date": "2019/06/17",
+    "numOunces": 49
+  },
+  {
+    "userID": 1,
+    "date": "2019/06/18",
+    "numOunces": 61
+  },
+  {
+    "userID": 2,
+    "date": "2019/06/18",
+    "numOunces": 70
+  },
+  {
+    "userID": 3,
+    "date": "2019/06/18",
+    "numOunces": 40
+  },
+  {
+    "userID": 4,
+    "date": "2019/06/18",
+    "numOunces": 93
+  },
+  {
+    "userID": 5,
+    "date": "2019/06/18",
+    "numOunces": 39
+  },
+  {
+    "userID": 6,
+    "date": "2019/06/18",
+    "numOunces": 61
+  },
+  {
+    "userID": 7,
+    "date": "2019/06/18",
+    "numOunces": 58
+  },
+  {
+    "userID": 8,
+    "date": "2019/06/18",
+    "numOunces": 90
+  },
+  {
+    "userID": 9,
+    "date": "2019/06/18",
+    "numOunces": 22
+  },
+  {
+    "userID": 10,
+    "date": "2019/06/18",
+    "numOunces": 97
+  },
+  {
+    "userID": 1,
+    "date": "2019/06/19",
+    "numOunces": 91
+  },
+  {
+    "userID": 2,
+    "date": "2019/06/19",
+    "numOunces": 76
+  },
+  {
+    "userID": 3,
+    "date": "2019/06/19",
+    "numOunces": 85
+  },
+  {
+    "userID": 4,
+    "date": "2019/06/19",
+    "numOunces": 21
+  },
+  {
+    "userID": 5,
+    "date": "2019/06/19",
+    "numOunces": 69
+  },
+  {
+    "userID": 6,
+    "date": "2019/06/19",
+    "numOunces": 41
+  },
+  {
+    "userID": 7,
+    "date": "2019/06/19",
+    "numOunces": 41
+  },
+  {
+    "userID": 8,
+    "date": "2019/06/19",
+    "numOunces": 79
+  },
+  {
+    "userID": 9,
+    "date": "2019/06/19",
+    "numOunces": 82
+  },
+  {
+    "userID": 10,
+    "date": "2019/06/19",
+    "numOunces": 38
+  },
+  {
+    "userID": 1,
+    "date": "2019/06/20",
+    "numOunces": 50
+  },
+  {
+    "userID": 2,
+    "date": "2019/06/20",
+    "numOunces": 71
+  },
+  {
+    "userID": 3,
+    "date": "2019/06/20",
+    "numOunces": 51
+  },
+  {
+    "userID": 4,
+    "date": "2019/06/20",
+    "numOunces": 95
+  },
+  {
+    "userID": 5,
+    "date": "2019/06/20",
+    "numOunces": 89
+  },
+  {
+    "userID": 6,
+    "date": "2019/06/20",
+    "numOunces": 85
+  },
+  {
+    "userID": 7,
+    "date": "2019/06/20",
+    "numOunces": 80
+  },
+  {
+    "userID": 8,
+    "date": "2019/06/20",
+    "numOunces": 89
+  },
+  {
+    "userID": 9,
+    "date": "2019/06/20",
+    "numOunces": 58
+  },
+  {
+    "userID": 10,
+    "date": "2019/06/20",
+    "numOunces": 78
+  },
+  {
+    "userID": 1,
+    "date": "2019/06/21",
+    "numOunces": 50
+  },
+  {
+    "userID": 2,
+    "date": "2019/06/21",
+    "numOunces": 27
+  },
+  {
+    "userID": 3,
+    "date": "2019/06/21",
+    "numOunces": 41
+  },
+  {
+    "userID": 4,
+    "date": "2019/06/21",
+    "numOunces": 91
+  },
+  {
+    "userID": 5,
+    "date": "2019/06/21",
+    "numOunces": 73
+  },
+  {
+    "userID": 6,
+    "date": "2019/06/21",
+    "numOunces": 90
+  },
+  {
+    "userID": 7,
+    "date": "2019/06/21",
+    "numOunces": 95
+  },
+  {
+    "userID": 8,
+    "date": "2019/06/21",
+    "numOunces": 37
+  },
+  {
+    "userID": 9,
+    "date": "2019/06/21",
+    "numOunces": 83
+  },
+  {
+    "userID": 10,
+    "date": "2019/06/21",
+    "numOunces": 95
+  },
+  {
+    "userID": 1,
+    "date": "2019/06/22",
+    "numOunces": 43
+  },
+  {
+    "userID": 2,
+    "date": "2019/06/22",
+    "numOunces": 58
+  },
+  {
+    "userID": 3,
+    "date": "2019/06/22",
+    "numOunces": 78
+  },
+  {
+    "userID": 4,
+    "date": "2019/06/22",
+    "numOunces": 34
+  },
+  {
+    "userID": 5,
+    "date": "2019/06/22",
+    "numOunces": 97
+  },
+  {
+    "userID": 6,
+    "date": "2019/06/22",
+    "numOunces": 73
+  },
+  {
+    "userID": 7,
+    "date": "2019/06/22",
+    "numOunces": 74
+  },
+  {
+    "userID": 8,
+    "date": "2019/06/22",
+    "numOunces": 40
+  },
+  {
+    "userID": 9,
+    "date": "2019/06/22",
+    "numOunces": 25
+  },
+  {
+    "userID": 10,
+    "date": "2019/06/22",
+    "numOunces": 72
+  },
+  {
+    "userID": 1,
+    "date": "2019/06/23",
+    "numOunces": 39
+  },
+  {
+    "userID": 2,
+    "date": "2019/06/23",
+    "numOunces": 44
+  },
+  {
+    "userID": 3,
+    "date": "2019/06/23",
+    "numOunces": 35
+  },
+  {
+    "userID": 4,
+    "date": "2019/06/23",
+    "numOunces": 62
+  },
+  {
+    "userID": 5,
+    "date": "2019/06/23",
+    "numOunces": 73
+  },
+  {
+    "userID": 6,
+    "date": "2019/06/23",
+    "numOunces": 71
+  },
+  {
+    "userID": 7,
+    "date": "2019/06/23",
+    "numOunces": 94
+  },
+  {
+    "userID": 8,
+    "date": "2019/06/23",
+    "numOunces": 38
+  },
+  {
+    "userID": 9,
+    "date": "2019/06/23",
+    "numOunces": 24
+  },
+  {
+    "userID": 10,
+    "date": "2019/06/23",
+    "numOunces": 81
+  },
+  {
+    "userID": 1,
+    "date": "2019/06/24",
+    "numOunces": 61
+  },
+  {
+    "userID": 2,
+    "date": "2019/06/24",
+    "numOunces": 33
+  },
+  {
+    "userID": 3,
+    "date": "2019/06/24",
+    "numOunces": 40
+  },
+  {
+    "userID": 4,
+    "date": "2019/06/24",
+    "numOunces": 66
+  },
+  {
+    "userID": 5,
+    "date": "2019/06/24",
+    "numOunces": 30
+  },
+  {
+    "userID": 6,
+    "date": "2019/06/24",
+    "numOunces": 45
+  },
+  {
+    "userID": 7,
+    "date": "2019/06/24",
+    "numOunces": 54
+  },
+  {
+    "userID": 8,
+    "date": "2019/06/24",
+    "numOunces": 58
+  },
+  {
+    "userID": 9,
+    "date": "2019/06/24",
+    "numOunces": 79
+  },
+  {
+    "userID": 10,
+    "date": "2019/06/24",
+    "numOunces": 72
+  },
+  {
+    "userID": 1,
+    "date": "2019/06/25",
+    "numOunces": 51
+  },
+  {
+    "userID": 2,
+    "date": "2019/06/25",
+    "numOunces": 67
+  },
+  {
+    "userID": 3,
+    "date": "2019/06/25",
+    "numOunces": 47
+  },
+  {
+    "userID": 4,
+    "date": "2019/06/25",
+    "numOunces": 27
+  },
+  {
+    "userID": 5,
+    "date": "2019/06/25",
+    "numOunces": 76
+  },
+  {
+    "userID": 6,
+    "date": "2019/06/25",
+    "numOunces": 63
+  },
+  {
+    "userID": 7,
+    "date": "2019/06/25",
+    "numOunces": 23
+  },
+  {
+    "userID": 8,
+    "date": "2019/06/25",
+    "numOunces": 74
+  },
+  {
+    "userID": 9,
+    "date": "2019/06/25",
+    "numOunces": 90
+  },
+  {
+    "userID": 10,
+    "date": "2019/06/25",
+    "numOunces": 70
+  },
+  {
+    "userID": 1,
+    "date": "2019/06/26",
+    "numOunces": 52
+  },
+  {
+    "userID": 2,
+    "date": "2019/06/26",
+    "numOunces": 27
+  },
+  {
+    "userID": 3,
+    "date": "2019/06/26",
+    "numOunces": 48
+  },
+  {
+    "userID": 4,
+    "date": "2019/06/26",
+    "numOunces": 44
+  },
+  {
+    "userID": 5,
+    "date": "2019/06/26",
+    "numOunces": 26
+  },
+  {
+    "userID": 6,
+    "date": "2019/06/26",
+    "numOunces": 44
+  },
+  {
+    "userID": 7,
+    "date": "2019/06/26",
+    "numOunces": 23
+  },
+  {
+    "userID": 8,
+    "date": "2019/06/26",
+    "numOunces": 72
+  },
+  {
+    "userID": 9,
+    "date": "2019/06/26",
+    "numOunces": 96
+  },
+  {
+    "userID": 10,
+    "date": "2019/06/26",
+    "numOunces": 91
+  },
+  {
+    "userID": 1,
+    "date": "2019/06/27",
+    "numOunces": 29
+  },
+  {
+    "userID": 2,
+    "date": "2019/06/27",
+    "numOunces": 70
+  },
+  {
+    "userID": 3,
+    "date": "2019/06/27",
+    "numOunces": 89
+  },
+  {
+    "userID": 4,
+    "date": "2019/06/27",
+    "numOunces": 83
+  },
+  {
+    "userID": 5,
+    "date": "2019/06/27",
+    "numOunces": 52
+  },
+  {
+    "userID": 6,
+    "date": "2019/06/27",
+    "numOunces": 30
+  },
+  {
+    "userID": 7,
+    "date": "2019/06/27",
+    "numOunces": 30
+  },
+  {
+    "userID": 8,
+    "date": "2019/06/27",
+    "numOunces": 65
+  },
+  {
+    "userID": 9,
+    "date": "2019/06/27",
+    "numOunces": 91
+  },
+  {
+    "userID": 10,
+    "date": "2019/06/27",
+    "numOunces": 29
+  },
+  {
+    "userID": 1,
+    "date": "2019/06/28",
+    "numOunces": 57
+  },
+  {
+    "userID": 2,
+    "date": "2019/06/28",
+    "numOunces": 56
+  },
+  {
+    "userID": 3,
+    "date": "2019/06/28",
+    "numOunces": 66
+  },
+  {
+    "userID": 4,
+    "date": "2019/06/28",
+    "numOunces": 47
+  },
+  {
+    "userID": 5,
+    "date": "2019/06/28",
+    "numOunces": 97
+  },
+  {
+    "userID": 6,
+    "date": "2019/06/28",
+    "numOunces": 70
+  },
+  {
+    "userID": 7,
+    "date": "2019/06/28",
+    "numOunces": 95
+  },
+  {
+    "userID": 8,
+    "date": "2019/06/28",
+    "numOunces": 36
+  },
+  {
+    "userID": 9,
+    "date": "2019/06/28",
+    "numOunces": 29
+  },
+  {
+    "userID": 10,
+    "date": "2019/06/28",
+    "numOunces": 40
+  },
+  {
+    "userID": 1,
+    "date": "2019/06/29",
+    "numOunces": 99
+  },
+  {
+    "userID": 2,
+    "date": "2019/06/29",
+    "numOunces": 35
+  },
+  {
+    "userID": 3,
+    "date": "2019/06/29",
+    "numOunces": 41
+  },
+  {
+    "userID": 4,
+    "date": "2019/06/29",
+    "numOunces": 68
+  },
+  {
+    "userID": 5,
+    "date": "2019/06/29",
+    "numOunces": 86
+  },
+  {
+    "userID": 6,
+    "date": "2019/06/29",
+    "numOunces": 70
+  },
+  {
+    "userID": 7,
+    "date": "2019/06/29",
+    "numOunces": 52
+  },
+  {
+    "userID": 8,
+    "date": "2019/06/29",
+    "numOunces": 53
+  },
+  {
+    "userID": 9,
+    "date": "2019/06/29",
+    "numOunces": 77
+  },
+  {
+    "userID": 10,
+    "date": "2019/06/29",
+    "numOunces": 54
+  }
+]
+
+if (typeof module !== 'undefined') {
+  module.exports = hydrationData;
+}
