--- data/test-data.js
@@ -1,2859 +1,0 @@
-
-const sleepData = [
-  {
-    "userID": 1,
-    "date": "2019/06/15",
-    "hoursSlept": 6.1,
-    "sleepQuality": 2.2
-  },
-  {
-    "userID": 2,
-    "date": "2019/06/15",
-    "hoursSlept": 7,
-    "sleepQuality": 4.7
-  },
-  {
-    "userID": 3,
-    "date": "2019/06/15",
-    "hoursSlept": 10.8,
-    "sleepQuality": 4.7
-  },
-  {
-    "userID": 4,
-    "date": "2019/06/15",
-    "hoursSlept": 5.4,
-    "sleepQuality": 3
-  },
-  {
-    "userID": 5,
-    "date": "2019/06/15",
-    "hoursSlept": 4.1,
-    "sleepQuality": 3.6
-  },
-  {
-    "userID": 6,
-    "date": "2019/06/15",
-    "hoursSlept": 9.6,
-    "sleepQuality": 2.9
-  },
-  {
-    "userID": 7,
-    "date": "2019/06/15",
-    "hoursSlept": 5.1,
-    "sleepQuality": 2.6
-  },
-  {
-    "userID": 8,
-    "date": "2019/06/15",
-    "hoursSlept": 8.1,
-    "sleepQuality": 3.5
-  },
-  {
-    "userID": 9,
-    "date": "2019/06/15",
-    "hoursSlept": 8.9,
-    "sleepQuality": 2.2
-  },
-  {
-    "userID": 10,
-    "date": "2019/06/15",
-    "hoursSlept": 4.4,
-    "sleepQuality": 1.6
-  },
-  {
-    "userID": 11,
-    "date": "2019/06/15",
-    "hoursSlept": 5,
-    "sleepQuality": 0
-  },
-  {
-    "userID": 1,
-    "date": "2019/06/16",
-    "hoursSlept": 4.1,
-    "sleepQuality": 3.8
-  },
-  {
-    "userID": 2,
-    "date": "2019/06/16",
-    "hoursSlept": 7.5,
-    "sleepQuality": 3.8
-  },
-  {
-    "userID": 3,
-    "date": "2019/06/16",
-    "hoursSlept": 10.7,
-    "sleepQuality": 3.4
-  },
-  {
-    "userID": 4,
-    "date": "2019/06/16",
-    "hoursSlept": 8.3,
-    "sleepQuality": 4.5
-  },
-  {
-    "userID": 5,
-    "date": "2019/06/16",
-    "hoursSlept": 7.4,
-    "sleepQuality": 2.4
-  },
-  {
-    "userID": 6,
-    "date": "2019/06/16",
-    "hoursSlept": 5.3,
-    "sleepQuality": 4.3
-  },
-  {
-    "userID": 7,
-    "date": "2019/06/16",
-    "hoursSlept": 10,
-    "sleepQuality": 4.5
-  },
-  {
-    "userID": 8,
-    "date": "2019/06/16",
-    "hoursSlept": 6.8,
-    "sleepQuality": 3
-  },
-  {
-    "userID": 9,
-    "date": "2019/06/16",
-    "hoursSlept": 9.1,
-    "sleepQuality": 1.2
-  },
-  {
-    "userID": 10,
-    "date": "2019/06/16",
-    "hoursSlept": 8,
-    "sleepQuality": 4.4
-  },
-  {
-    "userID": 1,
-    "date": "2019/06/17",
-    "hoursSlept": 8,
-    "sleepQuality": 2.6
-  },
-  {
-    "userID": 2,
-    "date": "2019/06/17",
-    "hoursSlept": 5.7,
-    "sleepQuality": 3
-  },
-  {
-    "userID": 3,
-    "date": "2019/06/17",
-    "hoursSlept": 5.3,
-    "sleepQuality": 4.9
-  },
-  {
-    "userID": 4,
-    "date": "2019/06/17",
-    "hoursSlept": 5.7,
-    "sleepQuality": 1.1
-  },
-  {
-    "userID": 5,
-    "date": "2019/06/17",
-    "hoursSlept": 10.5,
-    "sleepQuality": 3.7
-  },
-  {
-    "userID": 6,
-    "date": "2019/06/17",
-    "hoursSlept": 9.7,
-    "sleepQuality": 1.4
-  },
-  {
-    "userID": 7,
-    "date": "2019/06/17",
-    "hoursSlept": 9.9,
-    "sleepQuality": 4.1
-  },
-  {
-    "userID": 8,
-    "date": "2019/06/17",
-    "hoursSlept": 9.2,
-    "sleepQuality": 2.2
-  },
-  {
-    "userID": 9,
-    "date": "2019/06/17",
-    "hoursSlept": 7.2,
-    "sleepQuality": 2.2
-  },
-  {
-    "userID": 10,
-    "date": "2019/06/17",
-    "hoursSlept": 4.3,
-    "sleepQuality": 3.6
-  },
-  {
-    "userID": 1,
-    "date": "2019/06/18",
-    "hoursSlept": 10.4,
-    "sleepQuality": 3.1
-  },
-  {
-    "userID": 2,
-    "date": "2019/06/18",
-    "hoursSlept": 10.8,
-    "sleepQuality": 3.2
-  },
-  {
-    "userID": 3,
-    "date": "2019/06/18",
-    "hoursSlept": 9.8,
-    "sleepQuality": 2.6
-  },
-  {
-    "userID": 4,
-    "date": "2019/06/18",
-    "hoursSlept": 5.9,
-    "sleepQuality": 2.5
-  },
-  {
-    "userID": 5,
-    "date": "2019/06/18",
-    "hoursSlept": 5.2,
-    "sleepQuality": 4.1
-  },
-  {
-    "userID": 6,
-    "date": "2019/06/18",
-    "hoursSlept": 5.5,
-    "sleepQuality": 4
-  },
-  {
-    "userID": 7,
-    "date": "2019/06/18",
-    "hoursSlept": 5.5,
-    "sleepQuality": 1.4
-  },
-  {
-    "userID": 8,
-    "date": "2019/06/18",
-    "hoursSlept": 8.2,
-    "sleepQuality": 2.4
-  },
-  {
-    "userID": 9,
-    "date": "2019/06/18",
-    "hoursSlept": 5.5,
-    "sleepQuality": 2.4
-  },
-  {
-    "userID": 10,
-    "date": "2019/06/18",
-    "hoursSlept": 7,
-    "sleepQuality": 3.8
-  },
-  {
-    "userID": 1,
-    "date": "2019/06/19",
-    "hoursSlept": 10.7,
-    "sleepQuality": 1.2
-  },
-  {
-    "userID": 2,
-    "date": "2019/06/19",
-    "hoursSlept": 9.6,
-    "sleepQuality": 2.5
-  },
-  {
-    "userID": 3,
-    "date": "2019/06/19",
-    "hoursSlept": 7.2,
-    "sleepQuality": 3.4
-  },
-  {
-    "userID": 4,
-    "date": "2019/06/19",
-    "hoursSlept": 5.2,
-    "sleepQuality": 2.3
-  },
-  {
-    "userID": 5,
-    "date": "2019/06/19",
-    "hoursSlept": 4.8,
-    "sleepQuality": 3.4
-  },
-  {
-    "userID": 6,
-    "date": "2019/06/19",
-    "hoursSlept": 9.3,
-    "sleepQuality": 2
-  },
-  {
-    "userID": 7,
-    "date": "2019/06/19",
-    "hoursSlept": 5.1,
-    "sleepQuality": 4.4
-  },
-  {
-    "userID": 8,
-    "date": "2019/06/19",
-    "hoursSlept": 10.8,
-    "sleepQuality": 4.3
-  },
-  {
-    "userID": 9,
-    "date": "2019/06/19",
-    "hoursSlept": 10.4,
-    "sleepQuality": 2.6
-  },
-  {
-    "userID": 10,
-    "date": "2019/06/19",
-    "hoursSlept": 5.3,
-    "sleepQuality": 3.1
-  },
-  {
-    "userID": 1,
-    "date": "2019/06/20",
-    "hoursSlept": 9.3,
-    "sleepQuality": 1.2
-  },
-  {
-    "userID": 2,
-    "date": "2019/06/20",
-    "hoursSlept": 10.1,
-    "sleepQuality": 2.4
-  },
-  {
-    "userID": 3,
-    "date": "2019/06/20",
-    "hoursSlept": 9.4,
-    "sleepQuality": 1.2
-  },
-  {
-    "userID": 4,
-    "date": "2019/06/20",
-    "hoursSlept": 8.3,
-    "sleepQuality": 1.9
-  },
-  {
-    "userID": 5,
-    "date": "2019/06/20",
-    "hoursSlept": 10.1,
-    "sleepQuality": 3.5
-  },
-  {
-    "userID": 6,
-    "date": "2019/06/20",
-    "hoursSlept": 8.9,
-    "sleepQuality": 1.8
-  },
-  {
-    "userID": 7,
-    "date": "2019/06/20",
-    "hoursSlept": 7.1,
-    "sleepQuality": 4.9
-  },
-  {
-    "userID": 8,
-    "date": "2019/06/20",
-    "hoursSlept": 6.8,
-    "sleepQuality": 2.8
-  },
-  {
-    "userID": 9,
-    "date": "2019/06/20",
-    "hoursSlept": 9.8,
-    "sleepQuality": 3.9
-  },
-  {
-    "userID": 10,
-    "date": "2019/06/20",
-    "hoursSlept": 5.5,
-    "sleepQuality": 1.1
-  },
-  {
-    "userID": 1,
-    "date": "2019/06/21",
-    "hoursSlept": 7.8,
-    "sleepQuality": 4.2
-  },
-  {
-    "userID": 2,
-    "date": "2019/06/21",
-    "hoursSlept": 4.3,
-    "sleepQuality": 4.8
-  },
-  {
-    "userID": 3,
-    "date": "2019/06/21",
-    "hoursSlept": 8.9,
-    "sleepQuality": 3.7
-  },
-  {
-    "userID": 4,
-    "date": "2019/06/21",
-    "hoursSlept": 10.6,
-    "sleepQuality": 2.7
-  },
-  {
-    "userID": 5,
-    "date": "2019/06/21",
-    "hoursSlept": 9.6,
-    "sleepQuality": 4.1
-  },
-  {
-    "userID": 6,
-    "date": "2019/06/21",
-    "hoursSlept": 5.6,
-    "sleepQuality": 3.2
-  },
-  {
-    "userID": 7,
-    "date": "2019/06/21",
-    "hoursSlept": 7.5,
-    "sleepQuality": 3.3
-  },
-  {
-    "userID": 8,
-    "date": "2019/06/21",
-    "hoursSlept": 10.2,
-    "sleepQuality": 3
-  },
-  {
-    "userID": 9,
-    "date": "2019/06/21",
-    "hoursSlept": 7.7,
-    "sleepQuality": 2.9
-  },
-  {
-    "userID": 10,
-    "date": "2019/06/21",
-    "hoursSlept": 4.5,
-    "sleepQuality": 2.5
-  },
-  {
-    "userID": 1,
-    "date": "2019/06/22",
-    "hoursSlept": 7,
-    "sleepQuality": 3
-  },
-  {
-    "userID": 2,
-    "date": "2019/06/22",
-    "hoursSlept": 4.8,
-    "sleepQuality": 3.3
-  },
-  {
-    "userID": 3,
-    "date": "2019/06/22",
-    "hoursSlept": 9.8,
-    "sleepQuality": 2.1
-  },
-  {
-    "userID": 4,
-    "date": "2019/06/22",
-    "hoursSlept": 7.7,
-    "sleepQuality": 1.5
-  },
-  {
-    "userID": 5,
-    "date": "2019/06/22",
-    "hoursSlept": 8.1,
-    "sleepQuality": 1.3
-  },
-  {
-    "userID": 6,
-    "date": "2019/06/22",
-    "hoursSlept": 9.3,
-    "sleepQuality": 4.9
-  },
-  {
-    "userID": 7,
-    "date": "2019/06/22",
-    "hoursSlept": 7.5,
-    "sleepQuality": 4.3
-  },
-  {
-    "userID": 8,
-    "date": "2019/06/22",
-    "hoursSlept": 8.3,
-    "sleepQuality": 4.5
-  },
-  {
-    "userID": 9,
-    "date": "2019/06/22",
-    "hoursSlept": 7.2,
-    "sleepQuality": 2.5
-  },
-  {
-    "userID": 10,
-    "date": "2019/06/22",
-    "hoursSlept": 9,
-    "sleepQuality": 1.7
-  },
-  {
-    "userID": 1,
-    "date": "2019/06/23",
-    "hoursSlept": 7.8,
-    "sleepQuality": 1.5
-  },
-  {
-    "userID": 2,
-    "date": "2019/06/23",
-    "hoursSlept": 8,
-    "sleepQuality": 4.9
-  },
-  {
-    "userID": 3,
-    "date": "2019/06/23",
-    "hoursSlept": 4.7,
-    "sleepQuality": 3.9
-  },
-  {
-    "userID": 4,
-    "date": "2019/06/23",
-    "hoursSlept": 9.9,
-    "sleepQuality": 2.6
-  },
-  {
-    "userID": 5,
-    "date": "2019/06/23",
-    "hoursSlept": 5.1,
-    "sleepQuality": 1.6
-  },
-  {
-    "userID": 6,
-    "date": "2019/06/23",
-    "hoursSlept": 8.5,
-    "sleepQuality": 4.2
-  },
-  {
-    "userID": 7,
-    "date": "2019/06/23",
-    "hoursSlept": 9,
-    "sleepQuality": 3.6
-  },
-  {
-    "userID": 8,
-    "date": "2019/06/23",
-    "hoursSlept": 4.6,
-    "sleepQuality": 1.8
-  },
-  {
-    "userID": 9,
-    "date": "2019/06/23",
-    "hoursSlept": 4.1,
-    "sleepQuality": 3.8
-  },
-  {
-    "userID": 10,
-    "date": "2019/06/23",
-    "hoursSlept": 7.8,
-    "sleepQuality": 4.7
-  },
-  {
-    "userID": 1,
-    "date": "2019/06/24",
-    "hoursSlept": 8,
-    "sleepQuality": 1.3
-  },
-  {
-    "userID": 2,
-    "date": "2019/06/24",
-    "hoursSlept": 10.8,
-    "sleepQuality": 1
-  },
-  {
-    "userID": 3,
-    "date": "2019/06/24",
-    "hoursSlept": 9.3,
-    "sleepQuality": 1.8
-  },
-  {
-    "userID": 4,
-    "date": "2019/06/24",
-    "hoursSlept": 5,
-    "sleepQuality": 3.5
-  },
-  {
-    "userID": 5,
-    "date": "2019/06/24",
-    "hoursSlept": 6.8,
-    "sleepQuality": 2.1
-  },
-  {
-    "userID": 6,
-    "date": "2019/06/24",
-    "hoursSlept": 4.4,
-    "sleepQuality": 2.1
-  },
-  {
-    "userID": 7,
-    "date": "2019/06/24",
-    "hoursSlept": 10,
-    "sleepQuality": 2.4
-  },
-  {
-    "userID": 8,
-    "date": "2019/06/24",
-    "hoursSlept": 5.6,
-    "sleepQuality": 3.8
-  },
-  {
-    "userID": 9,
-    "date": "2019/06/24",
-    "hoursSlept": 7.2,
-    "sleepQuality": 1.4
-  },
-  {
-    "userID": 10,
-    "date": "2019/06/24",
-    "hoursSlept": 4.2,
-    "sleepQuality": 5
-  },
-  {
-    "userID": 1,
-    "date": "2019/06/25",
-    "hoursSlept": 5.1,
-    "sleepQuality": 3.7
-  },
-  {
-    "userID": 2,
-    "date": "2019/06/25",
-    "hoursSlept": 9.7,
-    "sleepQuality": 1.3
-  },
-  {
-    "userID": 3,
-    "date": "2019/06/25",
-    "hoursSlept": 6.4,
-    "sleepQuality": 4.9
-  },
-  {
-    "userID": 4,
-    "date": "2019/06/25",
-    "hoursSlept": 7.2,
-    "sleepQuality": 2.4
-  },
-  {
-    "userID": 5,
-    "date": "2019/06/25",
-    "hoursSlept": 6.3,
-    "sleepQuality": 4.4
-  },
-  {
-    "userID": 6,
-    "date": "2019/06/25",
-    "hoursSlept": 8.7,
-    "sleepQuality": 4.9
-  },
-  {
-    "userID": 7,
-    "date": "2019/06/25",
-    "hoursSlept": 8.4,
-    "sleepQuality": 2.9
-  },
-  {
-    "userID": 8,
-    "date": "2019/06/25",
-    "hoursSlept": 7.1,
-    "sleepQuality": 1.6
-  },
-  {
-    "userID": 9,
-    "date": "2019/06/25",
-    "hoursSlept": 4.5,
-    "sleepQuality": 1.1
-  },
-  {
-    "userID": 10,
-    "date": "2019/06/25",
-    "hoursSlept": 6.2,
-    "sleepQuality": 3.2
-  },
-  {
-    "userID": 1,
-    "date": "2019/06/26",
-    "hoursSlept": 7.7,
-    "sleepQuality": 2.4
-  },
-  {
-    "userID": 2,
-    "date": "2019/06/26",
-    "hoursSlept": 9.3,
-    "sleepQuality": 2.6
-  },
-  {
-    "userID": 3,
-    "date": "2019/06/26",
-    "hoursSlept": 6.1,
-    "sleepQuality": 1.5
-  },
-  {
-    "userID": 4,
-    "date": "2019/06/26",
-    "hoursSlept": 10.5,
-    "sleepQuality": 5
-  },
-  {
-    "userID": 5,
-    "date": "2019/06/26",
-    "hoursSlept": 9.9,
-    "sleepQuality": 1.8
-  },
-  {
-    "userID": 6,
-    "date": "2019/06/26",
-    "hoursSlept": 9,
-    "sleepQuality": 2
-  },
-  {
-    "userID": 7,
-    "date": "2019/06/26",
-    "hoursSlept": 9.7,
-    "sleepQuality": 2.8
-  },
-  {
-    "userID": 8,
-    "date": "2019/06/26",
-    "hoursSlept": 6.5,
-    "sleepQuality": 4.2
-  },
-  {
-    "userID": 9,
-    "date": "2019/06/26",
-    "hoursSlept": 4.3,
-    "sleepQuality": 1
-  },
-  {
-    "userID": 10,
-    "date": "2019/06/26",
-    "hoursSlept": 9.4,
-    "sleepQuality": 2.3
-  },
-  {
-    "userID": 1,
-    "date": "2019/06/27",
-    "hoursSlept": 9.4,
-    "sleepQuality": 4.6
-  },
-  {
-    "userID": 2,
-    "date": "2019/06/27",
-    "hoursSlept": 9,
-    "sleepQuality": 1.3
-  },
-  {
-    "userID": 3,
-    "date": "2019/06/27",
-    "hoursSlept": 4.3,
-    "sleepQuality": 4.2
-  },
-  {
-    "userID": 4,
-    "date": "2019/06/27",
-    "hoursSlept": 4.3,
-    "sleepQuality": 2.9
-  },
-  {
-    "userID": 5,
-    "date": "2019/06/27",
-    "hoursSlept": 4.5,
-    "sleepQuality": 3.9
-  },
-  {
-    "userID": 6,
-    "date": "2019/06/27",
-    "hoursSlept": 5.5,
-    "sleepQuality": 2
-  },
-  {
-    "userID": 7,
-    "date": "2019/06/27",
-    "hoursSlept": 5.1,
-    "sleepQuality": 2.8
-  },
-  {
-    "userID": 8,
-    "date": "2019/06/27",
-    "hoursSlept": 10.3,
-    "sleepQuality": 4.6
-  },
-  {
-    "userID": 9,
-    "date": "2019/06/27",
-    "hoursSlept": 5.2,
-    "sleepQuality": 2.5
-  },
-  {
-    "userID": 10,
-    "date": "2019/06/27",
-    "hoursSlept": 7.1,
-    "sleepQuality": 4.7
-  },
-  {
-    "userID": 1,
-    "date": "2019/06/28",
-    "hoursSlept": 7.6,
-    "sleepQuality": 4.7
-  },
-  {
-    "userID": 2,
-    "date": "2019/06/28",
-    "hoursSlept": 5.2,
-    "sleepQuality": 4.9
-  },
-  {
-    "userID": 3,
-    "date": "2019/06/28",
-    "hoursSlept": 4.3,
-    "sleepQuality": 3.7
-  },
-  {
-    "userID": 4,
-    "date": "2019/06/28",
-    "hoursSlept": 6.4,
-    "sleepQuality": 4.2
-  },
-  {
-    "userID": 5,
-    "date": "2019/06/28",
-    "hoursSlept": 5.5,
-    "sleepQuality": 1.5
-  },
-  {
-    "userID": 6,
-    "date": "2019/06/28",
-    "hoursSlept": 8.3,
-    "sleepQuality": 2.7
-  },
-  {
-    "userID": 7,
-    "date": "2019/06/28",
-    "hoursSlept": 5.9,
-    "sleepQuality": 2.4
-  },
-  {
-    "userID": 8,
-    "date": "2019/06/28",
-    "hoursSlept": 6.1,
-    "sleepQuality": 1.7
-  },
-  {
-    "userID": 9,
-    "date": "2019/06/28",
-    "hoursSlept": 7.2,
-    "sleepQuality": 1.4
-  },
-  {
-    "userID": 10,
-    "date": "2019/06/28",
-    "hoursSlept": 5.5,
-    "sleepQuality": 2
-  },
-  {
-    "userID": 1,
-    "date": "2019/06/29",
-    "hoursSlept": 5.3,
-    "sleepQuality": 1.2
-  },
-  {
-    "userID": 2,
-    "date": "2019/06/29",
-    "hoursSlept": 5.1,
-    "sleepQuality": 3.6
-  },
-  {
-    "userID": 3,
-    "date": "2019/06/29",
-    "hoursSlept": 6.6,
-    "sleepQuality": 4
-  },
-  {
-    "userID": 4,
-    "date": "2019/06/29",
-    "hoursSlept": 5.2,
-    "sleepQuality": 4.6
-  },
-  {
-    "userID": 5,
-    "date": "2019/06/29",
-    "hoursSlept": 7.3,
-    "sleepQuality": 4.8
-  },
-  {
-    "userID": 6,
-    "date": "2019/06/29",
-    "hoursSlept": 8.1,
-    "sleepQuality": 2.4
-  },
-  {
-    "userID": 7,
-    "date": "2019/06/29",
-    "hoursSlept": 7.1,
-    "sleepQuality": 1.8
-  },
-  {
-    "userID": 8,
-    "date": "2019/06/29",
-    "hoursSlept": 9.7,
-    "sleepQuality": 3.6
-  },
-  {
-    "userID": 9,
-    "date": "2019/06/29",
-    "hoursSlept": 9.7,
-    "sleepQuality": 3.5
-  },
-  {
-    "userID": 10,
-    "date": "2019/06/29",
-    "hoursSlept": 4.3,
-    "sleepQuality": 4.6
-  }
-]
-
-const userData = [
-  {
-    "id": 1,
-    "name": "Luisa Hane",
-    "address": "15195 Nakia Tunnel, Erdmanport VA 19901-1697",
-    "email": "Diana.Hayes1@hotmail.com",
-    "strideLength": 4.3,
-    "dailyStepGoal": 10000,
-    "friends": [
-      16,
-      4,
-      8
-    ]
-  },
-  {
-    "id": 2,
-    "name": "Jarvis Considine",
-    "address": "30086 Kathryn Port, Ciceroland NE 07273",
-    "email": "Dimitri.Bechtelar11@gmail.com",
-    "strideLength": 4.5,
-    "dailyStepGoal": 5000,
-    "friends": [
-      9,
-      18,
-      24,
-      19
-    ]
-  },
-  {
-    "id": 3,
-    "name": "Herminia Witting",
-    "address": "85823 Bosco Fork, East Oscarstad MI 85126-5660",
-    "email": "Elwin.Tromp@yahoo.com",
-    "strideLength": 4.4,
-    "dailyStepGoal": 5000,
-    "friends": [
-      19,
-      11,
-      42,
-      33
-    ]
-  },
-  {
-    "id": 4,
-    "name": "Mae Connelly",
-    "address": "28926 Schinner Islands, Turnermouth NE 23720-3230",
-    "email": "Marcos_Pollich@hotmail.com",
-    "strideLength": 3.1,
-    "dailyStepGoal": 4000,
-    "friends": [
-      48,
-      7,
-      44,
-      8
-    ]
-  },
-  {
-    "id": 5,
-    "name": "Erick Schaden",
-    "address": "514 Mayert Walk, Jordaneside SC 55023-6523",
-    "email": "Vanessa_Gerhold@gmail.com",
-    "strideLength": 3.1,
-    "dailyStepGoal": 8000,
-    "friends": [
-      13,
-      44,
-      49,
-      33,
-      10
-    ]
-  },
-  {
-    "id": 6,
-    "name": "Jerrold Bogisich",
-    "address": "8283 Carroll Harbor, Borerfort CT 69020-3448",
-    "email": "Mercedes_Zboncak53@yahoo.com",
-    "strideLength": 3.7,
-    "dailyStepGoal": 11000,
-    "friends": [
-      11,
-      48,
-      15
-    ]
-  },
-  {
-    "id": 7,
-    "name": "Breanne Fay",
-    "address": "834 Retta Knoll, Stantonland MA 71627-4121",
-    "email": "Dashawn28@gmail.com",
-    "strideLength": 2.9,
-    "dailyStepGoal": 8000,
-    "friends": [
-      12,
-      27,
-      22,
-      30
-    ]
-  },
-  {
-    "id": 8,
-    "name": "Laney Abshire",
-    "address": "86416 Koch Inlet, North Kaciefurt MA 80635",
-    "email": "Janice_Nitzsche2@yahoo.com",
-    "strideLength": 2.8,
-    "dailyStepGoal": 2000,
-    "friends": [
-      11,
-      41,
-      23,
-      49
-    ]
-  },
-  {
-    "id": 9,
-    "name": "Myron Schmitt",
-    "address": "85251 Martina Fields, West Aletha MD 00163-5315",
-    "email": "Gerard_Langosh22@hotmail.com",
-    "strideLength": 3.8,
-    "dailyStepGoal": 6000,
-    "friends": [
-      16,
-      26,
-      17
-    ]
-  },
-  {
-    "id": 10,
-    "name": "Roslyn Bernhard",
-    "address": "22859 Sean Meadow, Lake Dorthyfort MA 89560",
-    "email": "Hillary83@gmail.com",
-    "strideLength": 4.1,
-    "dailyStepGoal": 3000,
-    "friends": [
-      40,
-      33
-    ]
-  },
-]
-
-const hydrationData = [
-  {
-    "userID": 1,
-    "date": "2019/06/15",
-    "numOunces": 37
-  },
-  {
-    "userID": 2,
-    "date": "2019/06/15",
-    "numOunces": 75
-  },
-  {
-    "userID": 3,
-    "date": "2019/06/15",
-    "numOunces": 47
-  },
-  {
-    "userID": 4,
-    "date": "2019/06/15",
-    "numOunces": 85
-  },
-  {
-    "userID": 5,
-    "date": "2019/06/15",
-    "numOunces": 42
-  },
-  {
-    "userID": 6,
-    "date": "2019/06/15",
-    "numOunces": 87
-  },
-  {
-    "userID": 7,
-    "date": "2019/06/15",
-    "numOunces": 94
-  },
-  {
-    "userID": 8,
-    "date": "2019/06/15",
-    "numOunces": 84
-  },
-  {
-    "userID": 9,
-    "date": "2019/06/15",
-    "numOunces": 39
-  },
-  {
-    "userID": 10,
-    "date": "2019/06/15",
-    "numOunces": 75
-  },
-  {
-    "userID": 1,
-    "date": "2019/06/16",
-    "numOunces": 69
-  },
-  {
-    "userID": 2,
-    "date": "2019/06/16",
-    "numOunces": 91
-  },
-  {
-    "userID": 3,
-    "date": "2019/06/16",
-    "numOunces": 99
-  },
-  {
-    "userID": 4,
-    "date": "2019/06/16",
-    "numOunces": 95
-  },
-  {
-    "userID": 5,
-    "date": "2019/06/16",
-    "numOunces": 79
-  },
-  {
-    "userID": 6,
-    "date": "2019/06/16",
-    "numOunces": 57
-  },
-  {
-    "userID": 7,
-    "date": "2019/06/16",
-    "numOunces": 49
-  },
-  {
-    "userID": 8,
-    "date": "2019/06/16",
-    "numOunces": 26
-  },
-  {
-    "userID": 9,
-    "date": "2019/06/16",
-    "numOunces": 98
-  },
-  {
-    "userID": 10,
-    "date": "2019/06/16",
-    "numOunces": 68
-  },
-  {
-    "userID": 1,
-    "date": "2019/06/17",
-    "numOunces": 96
-  },
-  {
-    "userID": 2,
-    "date": "2019/06/17",
-    "numOunces": 96
-  },
-  {
-    "userID": 3,
-    "date": "2019/06/17",
-    "numOunces": 28
-  },
-  {
-    "userID": 4,
-    "date": "2019/06/17",
-    "numOunces": 82
-  },
-  {
-    "userID": 5,
-    "date": "2019/06/17",
-    "numOunces": 99
-  },
-  {
-    "userID": 6,
-    "date": "2019/06/17",
-    "numOunces": 57
-  },
-  {
-    "userID": 7,
-    "date": "2019/06/17",
-    "numOunces": 50
-  },
-  {
-    "userID": 8,
-    "date": "2019/06/17",
-    "numOunces": 47
-  },
-  {
-    "userID": 9,
-    "date": "2019/06/17",
-    "numOunces": 86
-  },
-  {
-    "userID": 10,
-    "date": "2019/06/17",
-    "numOunces": 49
-  },
-  {
-    "userID": 1,
-    "date": "2019/06/18",
-    "numOunces": 61
-  },
-  {
-    "userID": 2,
-    "date": "2019/06/18",
-    "numOunces": 70
-  },
-  {
-    "userID": 3,
-    "date": "2019/06/18",
-    "numOunces": 40
-  },
-  {
-    "userID": 4,
-    "date": "2019/06/18",
-    "numOunces": 93
-  },
-  {
-    "userID": 5,
-    "date": "2019/06/18",
-    "numOunces": 39
-  },
-  {
-    "userID": 6,
-    "date": "2019/06/18",
-    "numOunces": 61
-  },
-  {
-    "userID": 7,
-    "date": "2019/06/18",
-    "numOunces": 58
-  },
-  {
-    "userID": 8,
-    "date": "2019/06/18",
-    "numOunces": 90
-  },
-  {
-    "userID": 9,
-    "date": "2019/06/18",
-    "numOunces": 22
-  },
-  {
-    "userID": 10,
-    "date": "2019/06/18",
-    "numOunces": 97
-  },
-  {
-    "userID": 1,
-    "date": "2019/06/19",
-    "numOunces": 91
-  },
-  {
-    "userID": 2,
-    "date": "2019/06/19",
-    "numOunces": 76
-  },
-  {
-    "userID": 3,
-    "date": "2019/06/19",
-    "numOunces": 85
-  },
-  {
-    "userID": 4,
-    "date": "2019/06/19",
-    "numOunces": 21
-  },
-  {
-    "userID": 5,
-    "date": "2019/06/19",
-    "numOunces": 69
-  },
-  {
-    "userID": 6,
-    "date": "2019/06/19",
-    "numOunces": 41
-  },
-  {
-    "userID": 7,
-    "date": "2019/06/19",
-    "numOunces": 41
-  },
-  {
-    "userID": 8,
-    "date": "2019/06/19",
-    "numOunces": 79
-  },
-  {
-    "userID": 9,
-    "date": "2019/06/19",
-    "numOunces": 82
-  },
-  {
-    "userID": 10,
-    "date": "2019/06/19",
-    "numOunces": 38
-  },
-  {
-    "userID": 1,
-    "date": "2019/06/20",
-    "numOunces": 50
-  },
-  {
-    "userID": 2,
-    "date": "2019/06/20",
-    "numOunces": 71
-  },
-  {
-    "userID": 3,
-    "date": "2019/06/20",
-    "numOunces": 51
-  },
-  {
-    "userID": 4,
-    "date": "2019/06/20",
-    "numOunces": 95
-  },
-  {
-    "userID": 5,
-    "date": "2019/06/20",
-    "numOunces": 89
-  },
-  {
-    "userID": 6,
-    "date": "2019/06/20",
-    "numOunces": 85
-  },
-  {
-    "userID": 7,
-    "date": "2019/06/20",
-    "numOunces": 80
-  },
-  {
-    "userID": 8,
-    "date": "2019/06/20",
-    "numOunces": 89
-  },
-  {
-    "userID": 9,
-    "date": "2019/06/20",
-    "numOunces": 58
-  },
-  {
-    "userID": 10,
-    "date": "2019/06/20",
-    "numOunces": 78
-  },
-  {
-    "userID": 1,
-    "date": "2019/06/21",
-    "numOunces": 50
-  },
-  {
-    "userID": 2,
-    "date": "2019/06/21",
-    "numOunces": 27
-  },
-  {
-    "userID": 3,
-    "date": "2019/06/21",
-    "numOunces": 41
-  },
-  {
-    "userID": 4,
-    "date": "2019/06/21",
-    "numOunces": 91
-  },
-  {
-    "userID": 5,
-    "date": "2019/06/21",
-    "numOunces": 73
-  },
-  {
-    "userID": 6,
-    "date": "2019/06/21",
-    "numOunces": 90
-  },
-  {
-    "userID": 7,
-    "date": "2019/06/21",
-    "numOunces": 95
-  },
-  {
-    "userID": 8,
-    "date": "2019/06/21",
-    "numOunces": 37
-  },
-  {
-    "userID": 9,
-    "date": "2019/06/21",
-    "numOunces": 83
-  },
-  {
-    "userID": 10,
-    "date": "2019/06/21",
-    "numOunces": 95
-  },
-  {
-    "userID": 1,
-    "date": "2019/06/22",
-    "numOunces": 43
-  },
-  {
-    "userID": 2,
-    "date": "2019/06/22",
-    "numOunces": 58
-  },
-  {
-    "userID": 3,
-    "date": "2019/06/22",
-    "numOunces": 78
-  },
-  {
-    "userID": 4,
-    "date": "2019/06/22",
-    "numOunces": 34
-  },
-  {
-    "userID": 5,
-    "date": "2019/06/22",
-    "numOunces": 97
-  },
-  {
-    "userID": 6,
-    "date": "2019/06/22",
-    "numOunces": 73
-  },
-  {
-    "userID": 7,
-    "date": "2019/06/22",
-    "numOunces": 74
-  },
-  {
-    "userID": 8,
-    "date": "2019/06/22",
-    "numOunces": 40
-  },
-  {
-    "userID": 9,
-    "date": "2019/06/22",
-    "numOunces": 25
-  },
-  {
-    "userID": 10,
-    "date": "2019/06/22",
-    "numOunces": 72
-  },
-  {
-    "userID": 1,
-    "date": "2019/06/23",
-    "numOunces": 39
-  },
-  {
-    "userID": 2,
-    "date": "2019/06/23",
-    "numOunces": 44
-  },
-  {
-    "userID": 3,
-    "date": "2019/06/23",
-    "numOunces": 35
-  },
-  {
-    "userID": 4,
-    "date": "2019/06/23",
-    "numOunces": 62
-  },
-  {
-    "userID": 5,
-    "date": "2019/06/23",
-    "numOunces": 73
-  },
-  {
-    "userID": 6,
-    "date": "2019/06/23",
-    "numOunces": 71
-  },
-  {
-    "userID": 7,
-    "date": "2019/06/23",
-    "numOunces": 94
-  },
-  {
-    "userID": 8,
-    "date": "2019/06/23",
-    "numOunces": 38
-  },
-  {
-    "userID": 9,
-    "date": "2019/06/23",
-    "numOunces": 24
-  },
-  {
-    "userID": 10,
-    "date": "2019/06/23",
-    "numOunces": 81
-  },
-  {
-    "userID": 1,
-    "date": "2019/06/24",
-    "numOunces": 61
-  },
-  {
-    "userID": 2,
-    "date": "2019/06/24",
-    "numOunces": 33
-  },
-  {
-    "userID": 3,
-    "date": "2019/06/24",
-    "numOunces": 40
-  },
-  {
-    "userID": 4,
-    "date": "2019/06/24",
-    "numOunces": 66
-  },
-  {
-    "userID": 5,
-    "date": "2019/06/24",
-    "numOunces": 30
-  },
-  {
-    "userID": 6,
-    "date": "2019/06/24",
-    "numOunces": 45
-  },
-  {
-    "userID": 7,
-    "date": "2019/06/24",
-    "numOunces": 54
-  },
-  {
-    "userID": 8,
-    "date": "2019/06/24",
-    "numOunces": 58
-  },
-  {
-    "userID": 9,
-    "date": "2019/06/24",
-    "numOunces": 79
-  },
-  {
-    "userID": 10,
-    "date": "2019/06/24",
-    "numOunces": 72
-  },
-  {
-    "userID": 1,
-    "date": "2019/06/25",
-    "numOunces": 51
-  },
-  {
-    "userID": 2,
-    "date": "2019/06/25",
-    "numOunces": 67
-  },
-  {
-    "userID": 3,
-    "date": "2019/06/25",
-    "numOunces": 47
-  },
-  {
-    "userID": 4,
-    "date": "2019/06/25",
-    "numOunces": 27
-  },
-  {
-    "userID": 5,
-    "date": "2019/06/25",
-    "numOunces": 76
-  },
-  {
-    "userID": 6,
-    "date": "2019/06/25",
-    "numOunces": 63
-  },
-  {
-    "userID": 7,
-    "date": "2019/06/25",
-    "numOunces": 23
-  },
-  {
-    "userID": 8,
-    "date": "2019/06/25",
-    "numOunces": 74
-  },
-  {
-    "userID": 9,
-    "date": "2019/06/25",
-    "numOunces": 90
-  },
-  {
-    "userID": 10,
-    "date": "2019/06/25",
-    "numOunces": 70
-  },
-  {
-    "userID": 1,
-    "date": "2019/06/26",
-    "numOunces": 52
-  },
-  {
-    "userID": 2,
-    "date": "2019/06/26",
-    "numOunces": 27
-  },
-  {
-    "userID": 3,
-    "date": "2019/06/26",
-    "numOunces": 48
-  },
-  {
-    "userID": 4,
-    "date": "2019/06/26",
-    "numOunces": 44
-  },
-  {
-    "userID": 5,
-    "date": "2019/06/26",
-    "numOunces": 26
-  },
-  {
-    "userID": 6,
-    "date": "2019/06/26",
-    "numOunces": 44
-  },
-  {
-    "userID": 7,
-    "date": "2019/06/26",
-    "numOunces": 23
-  },
-  {
-    "userID": 8,
-    "date": "2019/06/26",
-    "numOunces": 72
-  },
-  {
-    "userID": 9,
-    "date": "2019/06/26",
-    "numOunces": 96
-  },
-  {
-    "userID": 10,
-    "date": "2019/06/26",
-    "numOunces": 91
-  },
-    {
-    "userID": 1,
-    "date": "2019/06/27",
-    "numOunces": 29
-  },
-  {
-    "userID": 2,
-    "date": "2019/06/27",
-    "numOunces": 70
-  },
-  {
-    "userID": 3,
-    "date": "2019/06/27",
-    "numOunces": 89
-  },
-  {
-    "userID": 4,
-    "date": "2019/06/27",
-    "numOunces": 83
-  },
-  {
-    "userID": 5,
-    "date": "2019/06/27",
-    "numOunces": 52
-  },
-  {
-    "userID": 6,
-    "date": "2019/06/27",
-    "numOunces": 30
-  },
-  {
-    "userID": 7,
-    "date": "2019/06/27",
-    "numOunces": 30
-  },
-  {
-    "userID": 8,
-    "date": "2019/06/27",
-    "numOunces": 65
-  },
-  {
-    "userID": 9,
-    "date": "2019/06/27",
-    "numOunces": 91
-  },
-  {
-    "userID": 10,
-    "date": "2019/06/27",
-    "numOunces": 29
-  },
-  {
-    "userID": 1,
-    "date": "2019/06/28",
-    "numOunces": 57
-  },
-  {
-    "userID": 2,
-    "date": "2019/06/28",
-    "numOunces": 56
-  },
-  {
-    "userID": 3,
-    "date": "2019/06/28",
-    "numOunces": 66
-  },
-  {
-    "userID": 4,
-    "date": "2019/06/28",
-    "numOunces": 47
-  },
-  {
-    "userID": 5,
-    "date": "2019/06/28",
-    "numOunces": 97
-  },
-  {
-    "userID": 6,
-    "date": "2019/06/28",
-    "numOunces": 70
-  },
-  {
-    "userID": 7,
-    "date": "2019/06/28",
-    "numOunces": 95
-  },
-  {
-    "userID": 8,
-    "date": "2019/06/28",
-    "numOunces": 36
-  },
-  {
-    "userID": 9,
-    "date": "2019/06/28",
-    "numOunces": 29
-  },
-  {
-    "userID": 10,
-    "date": "2019/06/28",
-    "numOunces": 40
-  },
-  {
-    "userID": 1,
-    "date": "2019/06/29",
-    "numOunces": 99
-  },
-  {
-    "userID": 2,
-    "date": "2019/06/29",
-    "numOunces": 35
-  },
-  {
-    "userID": 3,
-    "date": "2019/06/29",
-    "numOunces": 41
-  },
-  {
-    "userID": 4,
-    "date": "2019/06/29",
-    "numOunces": 68
-  },
-  {
-    "userID": 5,
-    "date": "2019/06/29",
-    "numOunces": 86
-  },
-  {
-    "userID": 6,
-    "date": "2019/06/29",
-    "numOunces": 70
-  },
-  {
-    "userID": 7,
-    "date": "2019/06/29",
-    "numOunces": 52
-  },
-  {
-    "userID": 8,
-    "date": "2019/06/29",
-    "numOunces": 53
-  },
-  {
-    "userID": 9,
-    "date": "2019/06/29",
-    "numOunces": 77
-  },
-  {
-    "userID": 10,
-    "date": "2019/06/29",
-    "numOunces": 54
-   }
-]
-
-const activityData = [
-  {
-    "userID": 1,
-    "date": "2019/06/15",
-    "numSteps": 3577,
-    "minutesActive": 140,
-    "flightsOfStairs": 16
-  },
-  {
-    "userID": 2,
-    "date": "2019/06/15",
-    "numSteps": 4294,
-    "minutesActive": 138,
-    "flightsOfStairs": 10
-  },
-  {
-    "userID": 3,
-    "date": "2019/06/15",
-    "numSteps": 7402,
-    "minutesActive": 116,
-    "flightsOfStairs": 33
-  },
-  {
-    "userID": 4,
-    "date": "2019/06/15",
-    "numSteps": 3486,
-    "minutesActive": 114,
-    "flightsOfStairs": 32
-  },
-  {
-    "userID": 5,
-    "date": "2019/06/15",
-    "numSteps": 11374,
-    "minutesActive": 213,
-    "flightsOfStairs": 13
-  },
-  {
-    "userID": 6,
-    "date": "2019/06/15",
-    "numSteps": 14810,
-    "minutesActive": 287,
-    "flightsOfStairs": 18
-  },
-  {
-    "userID": 7,
-    "date": "2019/06/15",
-    "numSteps": 2634,
-    "minutesActive": 107,
-    "flightsOfStairs": 5
-  },
-  {
-    "userID": 8,
-    "date": "2019/06/15",
-    "numSteps": 10333,
-    "minutesActive": 114,
-    "flightsOfStairs": 31
-  },
-  {
-    "userID": 9,
-    "date": "2019/06/15",
-    "numSteps": 6389,
-    "minutesActive": 41,
-    "flightsOfStairs": 33
-  },
-  {
-    "userID": 10,
-    "date": "2019/06/15",
-    "numSteps": 8015,
-    "minutesActive": 106,
-    "flightsOfStairs": 37
-  },
-  {
-    "userID": 1,
-    "date": "2019/06/16",
-    "numSteps": 6637,
-    "minutesActive": 175,
-    "flightsOfStairs": 36
-  },
-  {
-    "userID": 2,
-    "date": "2019/06/16",
-    "numSteps": 4112,
-    "minutesActive": 220,
-    "flightsOfStairs": 37
-  },
-  {
-    "userID": 3,
-    "date": "2019/06/16",
-    "numSteps": 12304,
-    "minutesActive": 152,
-    "flightsOfStairs": 8
-  },
-  {
-    "userID": 4,
-    "date": "2019/06/16",
-    "numSteps": 10689,
-    "minutesActive": 204,
-    "flightsOfStairs": 10
-  },
-  {
-    "userID": 5,
-    "date": "2019/06/16",
-    "numSteps": 9350,
-    "minutesActive": 167,
-    "flightsOfStairs": 38
-  },
-  {
-    "userID": 6,
-    "date": "2019/06/16",
-    "numSteps": 14461,
-    "minutesActive": 222,
-    "flightsOfStairs": 0
-  },
-  {
-    "userID": 7,
-    "date": "2019/06/16",
-    "numSteps": 4312,
-    "minutesActive": 40,
-    "flightsOfStairs": 38
-  },
-  {
-    "userID": 8,
-    "date": "2019/06/16",
-    "numSteps": 4217,
-    "minutesActive": 23,
-    "flightsOfStairs": 15
-  },
-  {
-    "userID": 9,
-    "date": "2019/06/16",
-    "numSteps": 4180,
-    "minutesActive": 49,
-    "flightsOfStairs": 28
-  },
-  {
-    "userID": 10,
-    "date": "2019/06/16",
-    "numSteps": 2895,
-    "minutesActive": 49,
-    "flightsOfStairs": 24
-  },
-  {
-    "userID": 1,
-    "date": "2019/06/17",
-    "numSteps": 14329,
-    "minutesActive": 168,
-    "flightsOfStairs": 18
-  },
-  {
-    "userID": 2,
-    "date": "2019/06/17",
-    "numSteps": 13750,
-    "minutesActive": 65,
-    "flightsOfStairs": 4
-  },
-  {
-    "userID": 3,
-    "date": "2019/06/17",
-    "numSteps": 4547,
-    "minutesActive": 97,
-    "flightsOfStairs": 5
-  },
-  {
-    "userID": 4,
-    "date": "2019/06/17",
-    "numSteps": 8160,
-    "minutesActive": 72,
-    "flightsOfStairs": 25
-  },
-  {
-    "userID": 5,
-    "date": "2019/06/17",
-    "numSteps": 13120,
-    "minutesActive": 151,
-    "flightsOfStairs": 25
-  },
-  {
-    "userID": 6,
-    "date": "2019/06/17",
-    "numSteps": 9463,
-    "minutesActive": 49,
-    "flightsOfStairs": 44
-  },
-  {
-    "userID": 7,
-    "date": "2019/06/17",
-    "numSteps": 2978,
-    "minutesActive": 152,
-    "flightsOfStairs": 18
-  },
-  {
-    "userID": 8,
-    "date": "2019/06/17",
-    "numSteps": 6233,
-    "minutesActive": 48,
-    "flightsOfStairs": 19
-  },
-  {
-    "userID": 9,
-    "date": "2019/06/17",
-    "numSteps": 7881,
-    "minutesActive": 39,
-    "flightsOfStairs": 4
-  },
-  {
-    "userID": 10,
-    "date": "2019/06/17",
-    "numSteps": 9052,
-    "minutesActive": 240,
-    "flightsOfStairs": 17
-  },
-  {
-    "userID": 1,
-    "date": "2019/06/18",
-    "numSteps": 4419,
-    "minutesActive": 165,
-    "flightsOfStairs": 33
-  },
-  {
-    "userID": 2,
-    "date": "2019/06/18",
-    "numSteps": 4662,
-    "minutesActive": 181,
-    "flightsOfStairs": 31
-  },
-  {
-    "userID": 3,
-    "date": "2019/06/18",
-    "numSteps": 2546,
-    "minutesActive": 274,
-    "flightsOfStairs": 26
-  },
-  {
-    "userID": 4,
-    "date": "2019/06/18",
-    "numSteps": 10056,
-    "minutesActive": 120,
-    "flightsOfStairs": 34
-  },
-  {
-    "userID": 5,
-    "date": "2019/06/18",
-    "numSteps": 4765,
-    "minutesActive": 35,
-    "flightsOfStairs": 25
-  },
-  {
-    "userID": 6,
-    "date": "2019/06/18",
-    "numSteps": 13658,
-    "minutesActive": 20,
-    "flightsOfStairs": 10
-  },
-  {
-    "userID": 7,
-    "date": "2019/06/18",
-    "numSteps": 10559,
-    "minutesActive": 200,
-    "flightsOfStairs": 32
-  },
-  {
-    "userID": 8,
-    "date": "2019/06/18",
-    "numSteps": 7658,
-    "minutesActive": 166,
-    "flightsOfStairs": 39
-  },
-  {
-    "userID": 9,
-    "date": "2019/06/18",
-    "numSteps": 12726,
-    "minutesActive": 188,
-    "flightsOfStairs": 39
-  },
-  {
-    "userID": 10,
-    "date": "2019/06/18",
-    "numSteps": 5113,
-    "minutesActive": 168,
-    "flightsOfStairs": 2
-  },
-  {
-    "userID": 1,
-    "date": "2019/06/19",
-    "numSteps": 8429,
-    "minutesActive": 275,
-    "flightsOfStairs": 2
-  },
-  {
-    "userID": 2,
-    "date": "2019/06/19",
-    "numSteps": 9858,
-    "minutesActive": 243,
-    "flightsOfStairs": 44
-  },
-  {
-    "userID": 3,
-    "date": "2019/06/19",
-    "numSteps": 10961,
-    "minutesActive": 188,
-    "flightsOfStairs": 17
-  },
-  {
-    "userID": 4,
-    "date": "2019/06/19",
-    "numSteps": 13451,
-    "minutesActive": 203,
-    "flightsOfStairs": 2
-  },
-  {
-    "userID": 5,
-    "date": "2019/06/19",
-    "numSteps": 8914,
-    "minutesActive": 191,
-    "flightsOfStairs": 6
-  },
-  {
-    "userID": 6,
-    "date": "2019/06/19",
-    "numSteps": 12737,
-    "minutesActive": 299,
-    "flightsOfStairs": 47
-  },
-  {
-    "userID": 7,
-    "date": "2019/06/19",
-    "numSteps": 5016,
-    "minutesActive": 183,
-    "flightsOfStairs": 41
-  },
-  {
-    "userID": 8,
-    "date": "2019/06/19",
-    "numSteps": 9775,
-    "minutesActive": 259,
-    "flightsOfStairs": 38
-  },
-  {
-    "userID": 9,
-    "date": "2019/06/19",
-    "numSteps": 5402,
-    "minutesActive": 187,
-    "flightsOfStairs": 33
-  },
-  {
-    "userID": 10,
-    "date": "2019/06/19",
-    "numSteps": 14645,
-    "minutesActive": 234,
-    "flightsOfStairs": 14
-  },
-  {
-    "userID": 1,
-    "date": "2019/06/20",
-    "numSteps": 14478,
-    "minutesActive": 140,
-    "flightsOfStairs": 12
-  },
-  {
-    "userID": 2,
-    "date": "2019/06/20",
-    "numSteps": 8153,
-    "minutesActive": 74,
-    "flightsOfStairs": 10
-  },
-  {
-    "userID": 3,
-    "date": "2019/06/20",
-    "numSteps": 5369,
-    "minutesActive": 129,
-    "flightsOfStairs": 46
-  },
-  {
-    "userID": 4,
-    "date": "2019/06/20",
-    "numSteps": 3314,
-    "minutesActive": 240,
-    "flightsOfStairs": 6
-  },
-  {
-    "userID": 5,
-    "date": "2019/06/20",
-    "numSteps": 4760,
-    "minutesActive": 219,
-    "flightsOfStairs": 46
-  },
-  {
-    "userID": 6,
-    "date": "2019/06/20",
-    "numSteps": 13316,
-    "minutesActive": 167,
-    "flightsOfStairs": 24
-  },
-  {
-    "userID": 7,
-    "date": "2019/06/20",
-    "numSteps": 13795,
-    "minutesActive": 286,
-    "flightsOfStairs": 31
-  },
-  {
-    "userID": 8,
-    "date": "2019/06/20",
-    "numSteps": 5680,
-    "minutesActive": 137,
-    "flightsOfStairs": 2
-  },
-  {
-    "userID": 9,
-    "date": "2019/06/20",
-    "numSteps": 14764,
-    "minutesActive": 125,
-    "flightsOfStairs": 30
-  },
-  {
-    "userID": 10,
-    "date": "2019/06/20",
-    "numSteps": 3487,
-    "minutesActive": 289,
-    "flightsOfStairs": 48
-  },
-  {
-    "userID": 1,
-    "date": "2019/06/21",
-    "numSteps": 6760,
-    "minutesActive": 135,
-    "flightsOfStairs": 6
-  },
-  {
-    "userID": 2,
-    "date": "2019/06/21",
-    "numSteps": 10225,
-    "minutesActive": 174,
-    "flightsOfStairs": 26
-  },
-  {
-    "userID": 3,
-    "date": "2019/06/21",
-    "numSteps": 7498,
-    "minutesActive": 199,
-    "flightsOfStairs": 13
-  },
-  {
-    "userID": 4,
-    "date": "2019/06/21",
-    "numSteps": 11807,
-    "minutesActive": 142,
-    "flightsOfStairs": 46
-  },
-  {
-    "userID": 5,
-    "date": "2019/06/21",
-    "numSteps": 2177,
-    "minutesActive": 142,
-    "flightsOfStairs": 21
-  },
-  {
-    "userID": 6,
-    "date": "2019/06/21",
-    "numSteps": 14091,
-    "minutesActive": 125,
-    "flightsOfStairs": 22
-  },
-  {
-    "userID": 7,
-    "date": "2019/06/21",
-    "numSteps": 13155,
-    "minutesActive": 179,
-    "flightsOfStairs": 46
-  },
-  {
-    "userID": 8,
-    "date": "2019/06/21",
-    "numSteps": 14267,
-    "minutesActive": 84,
-    "flightsOfStairs": 10
-  },
-  {
-    "userID": 9,
-    "date": "2019/06/21",
-    "numSteps": 9846,
-    "minutesActive": 281,
-    "flightsOfStairs": 14
-  },
-  {
-    "userID": 10,
-    "date": "2019/06/21",
-    "numSteps": 10854,
-    "minutesActive": 184,
-    "flightsOfStairs": 9
-  },
-  {
-    "userID": 1,
-    "date": "2019/06/22",
-    "numSteps": 10289,
-    "minutesActive": 119,
-    "flightsOfStairs": 6
-  },
-  {
-    "userID": 2,
-    "date": "2019/06/22",
-    "numSteps": 3605,
-    "minutesActive": 124,
-    "flightsOfStairs": 31
-  },
-  {
-    "userID": 3,
-    "date": "2019/06/22",
-    "numSteps": 11342,
-    "minutesActive": 53,
-    "flightsOfStairs": 17
-  },
-  {
-    "userID": 4,
-    "date": "2019/06/22",
-    "numSteps": 3595,
-    "minutesActive": 243,
-    "flightsOfStairs": 19
-  },
-  {
-    "userID": 5,
-    "date": "2019/06/22",
-    "numSteps": 10643,
-    "minutesActive": 213,
-    "flightsOfStairs": 37
-  },
-  {
-    "userID": 6,
-    "date": "2019/06/22",
-    "numSteps": 10833,
-    "minutesActive": 218,
-    "flightsOfStairs": 24
-  },
-  {
-    "userID": 7,
-    "date": "2019/06/22",
-    "numSteps": 13572,
-    "minutesActive": 280,
-    "flightsOfStairs": 32
-  },
-  {
-    "userID": 8,
-    "date": "2019/06/22",
-    "numSteps": 6894,
-    "minutesActive": 116,
-    "flightsOfStairs": 18
-  },
-  {
-    "userID": 9,
-    "date": "2019/06/22",
-    "numSteps": 3182,
-    "minutesActive": 222,
-    "flightsOfStairs": 15
-  },
-  {
-    "userID": 10,
-    "date": "2019/06/22",
-    "numSteps": 13309,
-    "minutesActive": 37,
-    "flightsOfStairs": 25
-  },
-  {
-    "userID": 1,
-    "date": "2019/06/23",
-    "numSteps": 13928,
-    "minutesActive": 218,
-    "flightsOfStairs": 21
-  },
-  {
-    "userID": 2,
-    "date": "2019/06/23",
-    "numSteps": 4148,
-    "minutesActive": 142,
-    "flightsOfStairs": 0
-  },
-  {
-    "userID": 3,
-    "date": "2019/06/23",
-    "numSteps": 4665,
-    "minutesActive": 219,
-    "flightsOfStairs": 9
-  },
-  {
-    "userID": 4,
-    "date": "2019/06/23",
-    "numSteps": 12134,
-    "minutesActive": 99,
-    "flightsOfStairs": 3
-  },
-  {
-    "userID": 5,
-    "date": "2019/06/23",
-    "numSteps": 7898,
-    "minutesActive": 190,
-    "flightsOfStairs": 23
-  },
-  {
-    "userID": 6,
-    "date": "2019/06/23",
-    "numSteps": 12967,
-    "minutesActive": 124,
-    "flightsOfStairs": 31
-  },
-  {
-    "userID": 7,
-    "date": "2019/06/23",
-    "numSteps": 10190,
-    "minutesActive": 79,
-    "flightsOfStairs": 8
-  },
-  {
-    "userID": 8,
-    "date": "2019/06/23",
-    "numSteps": 3337,
-    "minutesActive": 60,
-    "flightsOfStairs": 25
-  },
-  {
-    "userID": 9,
-    "date": "2019/06/23",
-    "numSteps": 6678,
-    "minutesActive": 136,
-    "flightsOfStairs": 19
-  },
-  {
-    "userID": 10,
-    "date": "2019/06/23",
-    "numSteps": 8735,
-    "minutesActive": 25,
-    "flightsOfStairs": 5
-  },
-  {
-    "userID": 1,
-    "date": "2019/06/24",
-    "numSteps": 7186,
-    "minutesActive": 25,
-    "flightsOfStairs": 15
-  },
-  {
-    "userID": 2,
-    "date": "2019/06/24",
-    "numSteps": 8568,
-    "minutesActive": 81,
-    "flightsOfStairs": 31
-  },
-  {
-    "userID": 3,
-    "date": "2019/06/24",
-    "numSteps": 10142,
-    "minutesActive": 269,
-    "flightsOfStairs": 45
-  },
-  {
-    "userID": 4,
-    "date": "2019/06/24",
-    "numSteps": 3704,
-    "minutesActive": 118,
-    "flightsOfStairs": 10
-  },
-  {
-    "userID": 5,
-    "date": "2019/06/24",
-    "numSteps": 4283,
-    "minutesActive": 123,
-    "flightsOfStairs": 2
-  },
-  {
-    "userID": 6,
-    "date": "2019/06/24",
-    "numSteps": 10188,
-    "minutesActive": 286,
-    "flightsOfStairs": 20
-  },
-  {
-    "userID": 7,
-    "date": "2019/06/24",
-    "numSteps": 10222,
-    "minutesActive": 24,
-    "flightsOfStairs": 10
-  },
-  {
-    "userID": 8,
-    "date": "2019/06/24",
-    "numSteps": 8190,
-    "minutesActive": 180,
-    "flightsOfStairs": 30
-  },
-  {
-    "userID": 9,
-    "date": "2019/06/24",
-    "numSteps": 14910,
-    "minutesActive": 126,
-    "flightsOfStairs": 3
-  },
-  {
-    "userID": 10,
-    "date": "2019/06/24",
-    "numSteps": 11738,
-    "minutesActive": 132,
-    "flightsOfStairs": 45
-  },
-  {
-    "userID": 1,
-    "date": "2019/06/25",
-    "numSteps": 3093,
-    "minutesActive": 185,
-    "flightsOfStairs": 25
-  },
-  {
-    "userID": 2,
-    "date": "2019/06/25",
-    "numSteps": 10305,
-    "minutesActive": 214,
-    "flightsOfStairs": 5
-  },
-  {
-    "userID": 3,
-    "date": "2019/06/25",
-    "numSteps": 4473,
-    "minutesActive": 158,
-    "flightsOfStairs": 37
-  },
-  {
-    "userID": 4,
-    "date": "2019/06/25",
-    "numSteps": 14547,
-    "minutesActive": 253,
-    "flightsOfStairs": 6
-  },
-  {
-    "userID": 5,
-    "date": "2019/06/25",
-    "numSteps": 4102,
-    "minutesActive": 204,
-    "flightsOfStairs": 21
-  },
-  {
-    "userID": 6,
-    "date": "2019/06/25",
-    "numSteps": 10314,
-    "minutesActive": 173,
-    "flightsOfStairs": 21
-  },
-  {
-    "userID": 7,
-    "date": "2019/06/25",
-    "numSteps": 4305,
-    "minutesActive": 206,
-    "flightsOfStairs": 26
-  },
-  {
-    "userID": 8,
-    "date": "2019/06/25",
-    "numSteps": 6862,
-    "minutesActive": 34,
-    "flightsOfStairs": 42
-  },
-  {
-    "userID": 9,
-    "date": "2019/06/25",
-    "numSteps": 9444,
-    "minutesActive": 169,
-    "flightsOfStairs": 10
-  },
-  {
-    "userID": 10,
-    "date": "2019/06/25",
-    "numSteps": 10987,
-    "minutesActive": 222,
-    "flightsOfStairs": 39
-  },
-  {
-    "userID": 1,
-    "date": "2019/06/26",
-    "numSteps": 8105,
-    "minutesActive": 219,
-    "flightsOfStairs": 28
-  },
-  {
-    "userID": 2,
-    "date": "2019/06/26",
-    "numSteps": 11522,
-    "minutesActive": 88,
-    "flightsOfStairs": 18
-  },
-  {
-    "userID": 3,
-    "date": "2019/06/26",
-    "numSteps": 3126,
-    "minutesActive": 229,
-    "flightsOfStairs": 45
-  },
-  {
-    "userID": 4,
-    "date": "2019/06/26",
-    "numSteps": 14142,
-    "minutesActive": 124,
-    "flightsOfStairs": 8
-  },
-  {
-    "userID": 5,
-    "date": "2019/06/26",
-    "numSteps": 11076,
-    "minutesActive": 26,
-    "flightsOfStairs": 25
-  },
-  {
-    "userID": 6,
-    "date": "2019/06/26",
-    "numSteps": 6394,
-    "minutesActive": 29,
-    "flightsOfStairs": 47
-  },
-  {
-    "userID": 7,
-    "date": "2019/06/26",
-    "numSteps": 11877,
-    "minutesActive": 169,
-    "flightsOfStairs": 9
-  },
-  {
-    "userID": 8,
-    "date": "2019/06/26",
-    "numSteps": 8161,
-    "minutesActive": 257,
-    "flightsOfStairs": 43
-  },
-  {
-    "userID": 9,
-    "date": "2019/06/26",
-    "numSteps": 7880,
-    "minutesActive": 98,
-    "flightsOfStairs": 20
-  },
-  {
-    "userID": 10,
-    "date": "2019/06/26",
-    "numSteps": 6549,
-    "minutesActive": 154,
-    "flightsOfStairs": 9
-  },
-  {
-    "userID": 1,
-    "date": "2019/06/27",
-    "numSteps": 3303,
-    "minutesActive": 79,
-    "flightsOfStairs": 39
-  },
-  {
-    "userID": 2,
-    "date": "2019/06/27",
-    "numSteps": 4240,
-    "minutesActive": 151,
-    "flightsOfStairs": 5
-  },
-  {
-    "userID": 3,
-    "date": "2019/06/27",
-    "numSteps": 9789,
-    "minutesActive": 265,
-    "flightsOfStairs": 45
-  },
-  {
-    "userID": 4,
-    "date": "2019/06/27",
-    "numSteps": 4053,
-    "minutesActive": 90,
-    "flightsOfStairs": 9
-  },
-  {
-    "userID": 5,
-    "date": "2019/06/27",
-    "numSteps": 10039,
-    "minutesActive": 122,
-    "flightsOfStairs": 49
-  },
-  {
-    "userID": 6,
-    "date": "2019/06/27",
-    "numSteps": 7765,
-    "minutesActive": 85,
-    "flightsOfStairs": 17
-  },
-  {
-    "userID": 7,
-    "date": "2019/06/27",
-    "numSteps": 3941,
-    "minutesActive": 251,
-    "flightsOfStairs": 27
-  },
-  {
-    "userID": 8,
-    "date": "2019/06/27",
-    "numSteps": 11565,
-    "minutesActive": 214,
-    "flightsOfStairs": 15
-  },
-  {
-    "userID": 9,
-    "date": "2019/06/27",
-    "numSteps": 11832,
-    "minutesActive": 206,
-    "flightsOfStairs": 34
-  },
-  {
-    "userID": 10,
-    "date": "2019/06/27",
-    "numSteps": 7115,
-    "minutesActive": 82,
-    "flightsOfStairs": 49
-  },
-  {
-    "userID": 1,
-    "date": "2019/06/28",
-    "numSteps": 10517,
-    "minutesActive": 169,
-    "flightsOfStairs": 6
-  },
-  {
-    "userID": 2,
-    "date": "2019/06/28",
-    "numSteps": 12555,
-    "minutesActive": 193,
-    "flightsOfStairs": 45
-  },
-  {
-    "userID": 3,
-    "date": "2019/06/28",
-    "numSteps": 6535,
-    "minutesActive": 188,
-    "flightsOfStairs": 19
-  },
-  {
-    "userID": 4,
-    "date": "2019/06/28",
-    "numSteps": 8401,
-    "minutesActive": 263,
-    "flightsOfStairs": 8
-  },
-  {
-    "userID": 5,
-    "date": "2019/06/28",
-    "numSteps": 14017,
-    "minutesActive": 118,
-    "flightsOfStairs": 29
-  },
-  {
-    "userID": 6,
-    "date": "2019/06/28",
-    "numSteps": 9596,
-    "minutesActive": 293,
-    "flightsOfStairs": 34
-  },
-  {
-    "userID": 7,
-    "date": "2019/06/28",
-    "numSteps": 8551,
-    "minutesActive": 63,
-    "flightsOfStairs": 32
-  },
-  {
-    "userID": 8,
-    "date": "2019/06/28",
-    "numSteps": 8192,
-    "minutesActive": 294,
-    "flightsOfStairs": 37
-  },
-  {
-    "userID": 9,
-    "date": "2019/06/28",
-    "numSteps": 9184,
-    "minutesActive": 95,
-    "flightsOfStairs": 49
-  },
-  {
-    "userID": 10,
-    "date": "2019/06/28",
-    "numSteps": 9288,
-    "minutesActive": 53,
-    "flightsOfStairs": 11
-  },
-  {
-    "userID": 1,
-    "date": "2019/06/29",
-    "numSteps": 6984,
-    "minutesActive": 33,
-    "flightsOfStairs": 6
-  },
-  {
-    "userID": 2,
-    "date": "2019/06/29",
-    "numSteps": 8096,
-    "minutesActive": 181,
-    "flightsOfStairs": 40
-  },
-  {
-    "userID": 3,
-    "date": "2019/06/29",
-    "numSteps": 11823,
-    "minutesActive": 85,
-    "flightsOfStairs": 5
-  },
-  {
-    "userID": 4,
-    "date": "2019/06/29",
-    "numSteps": 7702,
-    "minutesActive": 84,
-    "flightsOfStairs": 33
-  },
-  {
-    "userID": 5,
-    "date": "2019/06/29",
-    "numSteps": 4872,
-    "minutesActive": 285,
-    "flightsOfStairs": 11
-  },
-  {
-    "userID": 6,
-    "date": "2019/06/29",
-    "numSteps": 3026,
-    "minutesActive": 133,
-    "flightsOfStairs": 3
-  },
-  {
-    "userID": 7,
-    "date": "2019/06/29",
-    "numSteps": 11325,
-    "minutesActive": 118,
-    "flightsOfStairs": 23
-  },
-  {
-    "userID": 8,
-    "date": "2019/06/29",
-    "numSteps": 3512,
-    "minutesActive": 133,
-    "flightsOfStairs": 13
-  },
-  {
-    "userID": 9,
-    "date": "2019/06/29",
-    "numSteps": 9360,
-    "minutesActive": 140,
-    "flightsOfStairs": 7
-  },
-  {
-    "userID": 10,
-    "date": "2019/06/29",
-    "numSteps": 4943,
-    "minutesActive": 234,
-    "flightsOfStairs": 43
-  },
-
-]
-
-if (typeof module !== 'undefined') {
-  module.exports = sleepData, userData, hydrationData, activityData;
-}
