+++ test/UserRepository-test.js
@@ -3,51 +3,12 @@ const expect = chai.expect;
 const assert = require('chai').assert;
 
 const UserRepository = require('../src/UserRepository');
-// const userData = require('../data/users.js');
-// const testData = require('../data/test-data');
+const testUserData = require('../data/user-test-data.js')
 
 describe('User Repository', function() {
   let userRepository;
   beforeEach(() => {
-    testData = [
-      {
-        "id": 1,
-        "name": "Luisa Hane",
-        "address": "15195 Nakia Tunnel, Erdmanport VA 19901-1697",
-        "email": "Diana.Hayes1@hotmail.com",
-        "strideLength": 4.3,
-        "dailyStepGoal": 10000,
-        "friends": [
-          2,
-          3
-        ]
-      },
-      {
-        "id": 2,
-        "name": "Jarvis Considine",
-        "address": "30086 Kathryn Port, Ciceroland NE 07273",
-        "email": "Dimitri.Bechtelar11@gmail.com",
-        "strideLength": 4.5,
-        "dailyStepGoal": 5000,
-        "friends": [
-          1,
-          3
-        ]
-      },
-      {
-        "id": 3,
-        "name": "Herminia Witting",
-        "address": "85823 Bosco Fork, East Oscarstad MI 85126-5660",
-        "email": "Elwin.Tromp@yahoo.com",
-        "strideLength": 4.4,
-        "dailyStepGoal": 5000,
-        "friends": [
-          1,
-          2
-        ]
-      }
-    ]
-    userRepository = new UserRepository(testData);
+    userRepository = new UserRepository(testUserData);
   });
   
   it('should be a function', function() {
@@ -59,7 +20,7 @@ describe('User Repository', function() {
   })
 
   it('should have a parameter of user data', function() {
-    expect(userRepository.data).to.equal(testData);
+    expect(userRepository.data).to.equal(testUserData);
   })
 
   it('should not require a property to create a new User Repository', function () {
@@ -74,10 +35,10 @@ describe('User Repository', function() {
   })
 
   it('should find user data based on their ID', function() {
-    expect(userRepository.getDataById(1)).to.equal(testData[0]);
+    expect(userRepository.getDataById(1)).to.equal(testUserData[0]);
   })
 
   it('should average all users step goal', function() {
-    expect(userRepository.averageUserStepGoal()).to.equal(6667);
+    expect(userRepository.averageUserStepGoal()).to.equal(6200);
   })
 });
